Create `ForegroundNotificationManager` class
