+++ android/src/main/kotlin/net/mullvad/mullvadvpn/MullvadVpnService.kt
@@ -8,21 +8,14 @@ import kotlinx.coroutines.CompletableDeferred
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 
-import android.app.Activity
-import android.app.Notification
-import android.app.PendingIntent
-import android.content.Context
 import android.content.Intent
 import android.net.VpnService
 import android.os.Binder
 import android.os.IBinder
-import android.support.v4.app.NotificationCompat
 
 import net.mullvad.mullvadvpn.dataproxy.AppVersionInfoFetcher
 import net.mullvad.mullvadvpn.dataproxy.ConnectionProxy
 import net.mullvad.mullvadvpn.model.TunConfig
-
-val ONGOING_NOTIFICATION_ID: Int = 1
 
 class MullvadVpnService : VpnService() {
     private val created = CompletableDeferred<Unit>()
@@ -32,10 +25,11 @@ class MullvadVpnService : VpnService() {
 
     val daemon = startDaemon()
     val connectionProxy = ConnectionProxy(this, daemon)
+    val notificationManager = ForegroundNotificationManager(this)
 
     override fun onCreate() {
         versionInfoFetcher = AppVersionInfoFetcher(daemon, this)
-        startForeground(ONGOING_NOTIFICATION_ID, buildNotification())
+        notificationManager.onCreate()
         created.complete(Unit)
     }
 
@@ -45,10 +39,10 @@ class MullvadVpnService : VpnService() {
 
     override fun onDestroy() {
         connectionProxy.onDestroy()
+        notificationManager.onDestroy()
         versionInfoFetcher.stop()
         daemon.cancel()
         created.cancel()
-        stopForeground(ONGOING_NOTIFICATION_ID)
     }
 
     fun createTun(config: TunConfig): Int {
@@ -92,20 +86,6 @@ class MullvadVpnService : VpnService() {
 
             stopSelf()
         }
-    }
-
-    private fun buildNotification(): Notification {
-        val intent = Intent(this, MainActivity::class.java)
-            .setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP or Intent.FLAG_ACTIVITY_SINGLE_TOP)
-            .setAction(Intent.ACTION_MAIN)
-
-        val pendingIntent =
-            PendingIntent.getActivity(this, 1, intent, PendingIntent.FLAG_UPDATE_CURRENT)
-
-        return NotificationCompat.Builder(this)
-            .setSmallIcon(R.drawable.notification)
-            .setContentIntent(pendingIntent)
-            .build()
     }
 
     private fun startDaemon() = GlobalScope.async(Dispatchers.Default) {
